requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
openai>=1.30.0
numpy>=1.21.0
google-generativeai>=0.7.0
flask>=2.3.0
